+++ build.sh
@@ -1,0 +1,4 @@
+#!/bin/bash -ex
+
+VERSION=$(< VERSION)
+docker build --build-arg VERSION=$VERSION -t debify:$VERSION .