+++ ci/test.sh
@@ -1,0 +1,8 @@
+#!/bin/bash -ex
+
+bundle
+
+for target in spec cucumber; do
+  bundle exec rake $target || true
+done
+