+++ push-image.sh
@@ -1,0 +1,6 @@
+#!/bin/bash -ex
+
+TAG=$(< VERSION)
+
+docker push registry.tld/conjurinc/debify:$TAG
+docker push registry.tld/conjurinc/debify:latest